Rename `datamodules` folder to `data` (#501)
