+++ .gitignore
@@ -146,7 +146,6 @@ dmypy.json
 
 # Lightning-Hydra-Template
 configs/local/default.yaml
-data/
-logs/
+/data/
+/logs/
 .env
-.autoenv
