+++ configs/data/mnist.yaml
@@ -1,4 +1,4 @@
-_target_: src.datamodules.mnist_datamodule.MNISTDataModule
+_target_: src.data.mnist_datamodule.MNISTDataModule
 data_dir: ${paths.data_dir}
 batch_size: 128
 train_val_test_split: [55_000, 5_000, 10_000]
