+++ configs/eval.yaml
@@ -2,7 +2,7 @@
 
 defaults:
   - _self_
-  - datamodule: mnist.yaml # choose datamodule with `test_dataloader()` for evaluation
+  - data: mnist.yaml # choose datamodule with `test_dataloader()` for evaluation
   - model: mnist.yaml
   - logger: null
   - trainer: default.yaml
