+++ configs/hparams_search/mnist_optuna.yaml
@@ -46,7 +46,7 @@ hydra:
     # define hyperparameter search space
     params:
       model.optimizer.lr: interval(0.0001, 0.1)
-      datamodule.batch_size: choice(32, 64, 128, 256)
+      data.batch_size: choice(32, 64, 128, 256)
       model.net.lin1_size: choice(64, 128, 256)
       model.net.lin2_size: choice(64, 128, 256)
       model.net.lin3_size: choice(32, 64, 128, 256)
