+++ src/utils/rich_utils.py
@@ -18,7 +18,7 @@ log = pylogger.get_pylogger(__name__)
 def print_config_tree(
     cfg: DictConfig,
     print_order: Sequence[str] = (
-        "datamodule",
+        "data",
         "model",
         "callbacks",
         "logger",
