+++ tests/conftest.py
@@ -21,8 +21,8 @@ def cfg_train_global() -> DictConfig:
             cfg.trainer.limit_test_batches = 0.1
             cfg.trainer.accelerator = "cpu"
             cfg.trainer.devices = 1
-            cfg.datamodule.num_workers = 0
-            cfg.datamodule.pin_memory = False
+            cfg.data.num_workers = 0
+            cfg.data.pin_memory = False
             cfg.extras.print_config = False
             cfg.extras.enforce_tags = False
             cfg.logger = None
@@ -42,8 +42,8 @@ def cfg_eval_global() -> DictConfig:
             cfg.trainer.limit_test_batches = 0.1
             cfg.trainer.accelerator = "cpu"
             cfg.trainer.devices = 1
-            cfg.datamodule.num_workers = 0
-            cfg.datamodule.pin_memory = False
+            cfg.data.num_workers = 0
+            cfg.data.pin_memory = False
             cfg.extras.print_config = False
             cfg.extras.enforce_tags = False
             cfg.logger = None
