+++ tests/test_configs.py
@@ -5,25 +5,25 @@ from omegaconf import DictConfig
 
 def test_train_config(cfg_train: DictConfig):
     assert cfg_train
-    assert cfg_train.datamodule
+    assert cfg_train.data
     assert cfg_train.model
     assert cfg_train.trainer
 
     HydraConfig().set_config(cfg_train)
 
-    hydra.utils.instantiate(cfg_train.datamodule)
+    hydra.utils.instantiate(cfg_train.data)
     hydra.utils.instantiate(cfg_train.model)
     hydra.utils.instantiate(cfg_train.trainer)
 
 
 def test_eval_config(cfg_eval: DictConfig):
     assert cfg_eval
-    assert cfg_eval.datamodule
+    assert cfg_eval.data
     assert cfg_eval.model
     assert cfg_eval.trainer
 
     HydraConfig().set_config(cfg_eval)
 
-    hydra.utils.instantiate(cfg_eval.datamodule)
+    hydra.utils.instantiate(cfg_eval.data)
     hydra.utils.instantiate(cfg_eval.model)
     hydra.utils.instantiate(cfg_eval.trainer)
